refactor: add `configureProject` function from lines in `main` function
